+++ .gitignore
@@ -6,3 +6,7 @@ data/deneb_goerli_slot_7416760.json
 data/deneb_goerli_slot_7413760.json
 data/deneb_goerli_slot_7421952.json
 data/deneb_goerli_slot_7431952.json
+
+generation/proofGenCLI
+
+solidityProofGen/solidityProofGen
